Crate re_query2

Provide query-centric access to the [re_data_store].

Re-exports

pub use self::clamped_zip::*;

pub use self::range_zip::*;

Modules

clamped_zip

latest_at 🔒

promise 🔒

range 🔒

range_zip

visible_history 🔒

Structs

ComponentNotFoundError

ExtraQueryHistory

When showing an entity in the history view, add this much history to it.

LatestAtComponentResults

Uncached results for a particular component when using a latest-at query.

LatestAtResults

Raw results for a latest-at query.

Promise

A Promise turns a source DataCell into a new DataCell with the helper of a PromiseResolver.

PromiseId

Uniquely identifies a Promise.

PromiseResolver

Resolves and keeps track of Promises.

RangeComponentResults

Uncached results for a particular component when using a range query.

RangeResults

Raw results for a range query.

VisibleHistory

Visible history bounds.

Enums

PromiseResult

The result of resolving a Promise through a PromiseResolver.

QueryError

VisibleHistoryBoundary

One of the boundaries of the visible history.

Traits

ToArchetype

Helper extension trait to convert query results into re_types_core::Archetypes.

Functions

latest_at

Queries for the given component_names using latest-at semantics.

range

Queries for the given component_names using range semantics.

Type Aliases

Result